% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R, R/extendr-wrappers.R
\name{prql_to_sql}
\alias{prql_to_sql}
\alias{compile}
\title{Compile a PRQL query into a SQL query}
\usage{
prql_to_sql(prql)

compile(prql_query, dialect, format, signature_comment)
}
\arguments{
\item{prql_query}{a PRQL query string.}

\item{dialect}{a SQL dialect name to use. If it is not a valid value, the dialect contained in the query will be used.}

\item{format}{a logical flag. Whether to format the SQL query.}

\item{signature_comment}{a logical flag. Whether to add a signature comment to the output SQL query.}
}
\value{
a SQL query string
}
\description{
Compile a PRQL query into a SQL query
}
