% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{format_prql}
\alias{format_prql}
\title{Format a PRQL string}
\usage{
format_prql(prql)
}
\arguments{
\item{prql}{A PRQL string}
}
\value{
A PRQL string
}
\description{
Format a PRQL string
}
\examples{
format_prql("from mtcars | select cyl")

"
from mtcars
filter cyl > 6
select [cyl, mpg]
" |>
  format_prql() |>
  cat()
}
