% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{json_to_prql}
\alias{json_to_prql}
\title{Convert a JSON AST back to a PRQL string}
\usage{
json_to_prql(json)
}
\arguments{
\item{json}{A JSON string of AST}
}
\value{
A PRQL string
}
\description{
Convert a JSON AST back to a PRQL string
}
\examples{
"from mtcars | filter cyl > 6 | select [cyl, mpg]" |>
  prql_to_json() |>
  json_to_prql()
}
\seealso{
\code{\link[=prql_to_json]{prql_to_json()}}
}
