% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{prql_available_dialects}
\alias{prql_available_dialects}
\title{Available dialect names}
\usage{
prql_available_dialects()
}
\value{
a character vector of dialect names.
}
\description{
Available dialects for the \code{dialect} option of the \code{\link[=prql_compile]{prql_compile()}} function.
}
\examples{
prql_available_dialects()
}
