% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{prql_available_targets}
\alias{prql_available_targets}
\title{Available target names}
\usage{
prql_available_targets()
}
\value{
a character vector of target names.
}
\description{
Available targets for the \code{target} option of the \code{\link[=prql_compile]{prql_compile()}} function.
}
\examples{
prql_available_targets()
}
