% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{prql_compiler_version}
\alias{prql_compiler_version}
\title{prql-compiler's version}
\usage{
prql_compiler_version()
}
\value{
a \link{numeric_version} with the version of the prql-compiler.
}
\description{
prql-compiler's version
}
\examples{
prql_compiler_version()
}
