% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{prql_to_json}
\alias{prql_to_json}
\title{Compile a PRQL string into a JSON version of the Query}
\usage{
prql_to_json(prql)
}
\arguments{
\item{prql}{A PRQL string}
}
\value{
A JSON string of AST
}
\description{
Compile a PRQL string into a JSON version of the Query
}
\examples{
"from mtcars | filter cyl > 6 | select [cyl, mpg]" |>
  prql_to_json()
}
\seealso{
\code{\link[=json_to_prql]{json_to_prql()}}
}
