% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{prql_to_sql}
\alias{prql_to_sql}
\title{Compile a PRQL string into a SQL string}
\usage{
prql_to_sql(prql)
}
\arguments{
\item{prql}{A PRQL string}
}
\value{
A SQL string
}
\description{
Compile a PRQL string into a SQL string
}
\examples{
"from mtcars | filter cyl > 6 | select [cyl, mpg]" |>
  prql_to_sql()
}
