% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{prql_version}
\alias{prql_version}
\title{prqlc's version}
\usage{
prql_version()
}
\value{
a \link{numeric_version} with the version of the built-in prqlc.
}
\description{
prqlc's version
}
\examples{
prql_version()
}
