% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prqlr-package.R
\docType{package}
\name{prqlr-package}
\alias{prqlr}
\alias{prqlr-package}
\title{prqlr: R Bindings for the 'prql-compiler' Rust Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a function to convert 'PRQL' strings to 'SQL' strings. Combined with other R functions that take 'SQL' as an argument, 'PRQL' can be used on R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eitsupi.github.io/prqlr/}
  \item \url{https://github.com/eitsupi/prqlr}
  \item Report bugs at \url{https://github.com/eitsupi/prqlr/issues}
}

}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}

Authors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details)
}

}
\keyword{internal}
