% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwrap}
\alias{unwrap}
\title{rust-like unwrapping of result. Useful to do error handling on the R side.}
\usage{
unwrap(result, call = sys.call(1L))
}
\arguments{
\item{result}{a list where either element ok or err is NULL, or both if ok is a literal NULL.}

\item{call}{context of error or string used for error printing, could be omitted.}
}
\value{
the ok-element of list , or a error will be thrown
}
\description{
rust-like unwrapping of result. Useful to do error handling on the R side.
}
\examples{
#export unwrap for this example, only for internal use
unwrap <- prqlr:::unwrap

#unwrap a result with the ok-value "foo"
unwrap(list(ok="foo",err=NULL))

#unwrap an err-result
tryCatch(
  unwrap(ok=NULL, err = "something happend on the rust side"),
  error = function(e) as.character(e)
)
}
