% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{format_prql}
\alias{format_prql}
\title{Format a PRQL string}
\usage{
format_prql(prql)
}
\arguments{
\item{prql}{A PRQL string}
}
\value{
A PRQL string
}
\description{
Format a PRQL string
}
\examples{
format_prql("from mtcars | select cyl")

# Show the PQL string with the cli package's cli_code function
if (requireNamespace("cli", quietly = TRUE)) {
  library(cli)
  "
  from mtcars
  filter cyl > 6
  select [cyl, mpg]
  " |>
    format_prql() |>
    cli_code(language = "elm")
    # Coincidentally, elm's syntax highlighting is best suited for prql
}
}
