% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{prql_get_targets}
\alias{prql_get_targets}
\title{Get available target names}
\usage{
prql_get_targets()
}
\value{
a character vector of target names.
}
\description{
Get available target names for the \code{target} option of the \code{\link[=prql_compile]{prql_compile()}} function.
}
\examples{
prql_get_targets()
}
