% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{prql_version}
\alias{prql_version}
\title{prql-compiler's version}
\usage{
prql_version()
}
\value{
a \link{numeric_version} with the version of the built-in prql-compiler.
}
\description{
prql-compiler's version
}
\examples{
prql_version()
}
